==> R/00_all.qmd <==
---
title: "00_all"
format:
  html:
    self-contained: true
execute:
  echo: true
---

# Master Document - Breast Cancer Response Analysis

## Load Data

```{r child="01_load.qmd"}
```

## Clean Data

```{r child="02_clean.qmd"}
```

## Augment Data

```{r child="03_augment.qmd"}
```

## Describe Data

```{r child="04_describe.qmd"}
```

## Analysis

```{r child ="05_analysis_1.qmd"}
```

```{r child ="06_analysis_2.qmd"}
```

```{r child ="07_analysis_3.qmd"}
```

```{r child ="08_analysis_4.qmd"}
```

```{r child ="09_analysis_5.qmd"}
```

```{r child ="10_analysis_6.qmd"}
```

```{r child ="11_analysis_7.qmd"}
```

```{r child ="12_analysis_8.qmd"}
```

```{r child ="13_analysis_9.qmd"}
```

```{r child ="14_analysis_10.qmd"}
```

```{r child ="15_analysis_11.qmd"}
```

```{r child ="16_analysis_12.qmd"}
```

```{r child ="17_analysis_13.qmd"}
```

```{r child ="18_analysis_14.qmd"}
```

```{r child ="19_analysis_15.qmd"}
```

```{r child ="20_analysis_16.qmd"}
```

```{r child ="21_analysis_17.qmd"}
```

```{r child ="22_analysis_18.qmd"}
```

```{r child ="23_analysis_19.qmd"}
```

Table of Contents of Analysis files:

-   01 - E and P receptor statuses vs cancer aggressiveness

-   02 - HER2 receptor status vs cancer aggressiveness

-   03 - HER2 receptor status vs tumor size before treatment

-   04 - HER2 receptor status vs nodes before treatment

-   05 - E and P receptor statuses vs treatment outcome

-   06 - HER2 receptor status vs treatment outcome

-   07 - Cancer aggressiveness vs treatment outcome

-   08 - Age vs. Outcome

-   09 - Age and Race vs. Aggressivness and Tumor Size

-   10 - Tumorsize, Lymphnodes and Aggression Comparison

-   11 - Response to Treatment according to Receptor Status

-   12 - Her2 fish vs. Her2 receptor positive or negative and linear regression of her2 and tumor size

-   13 - PCA with colouring according to receptor positiveness or negativeness

-   14 - Response (pCR vs RD) by ER/PR/HER2 Genotypes

-   15 - Tumor Grade (1–3) by ER/PR/HER2 Genotypes

-   16 - Tumor Size Before Treatment by ER/PR/HER2 Genotypes

-   17 - Lymph Node Status Before Treatment by ER/PR/HER2 Genotypes

-   18 - Treatment Distribution Across ER/PR/HER2 Genotypes

-   19 - Age and Genotype Comparison
