==> R/01_load.qmd <==
---
title: "01_load"
format: 
  html:
    self-contained: true
editor: visual
---

## Load libraries

```{r}
#| message: false
library("GEOquery")
library("readr")
library("here")
library("tidyverse")
```

## Save raw data

```{r}
gse <- getGEO("GSE20194",
              GSEMatrix = TRUE,
              getGPL = FALSE)

save(gse,
     file = here("data/_raw/raw_data.RData"))
```

## Load raw data

```{r}
load(here("data/_raw/raw_data.RData"))

eset <- gse[[1]]
```

## Preliminary processing

Convert the ExpressionSet object to a data frame and handle missing or blank values to prevent misalignment.

```{r}
# Extract sample metadata and add a sample column
# Using %>% instead of |> because %>% allows using '.' as placeholder
dat_load <- as.data.frame(pData(eset))  %>% 
  mutate(
    sample_id = rownames(.)
  )  %>% 
  relocate(sample_id)  %>% 
  mutate(
    across(where(is.factor), as.character)
  )  %>%   
  mutate(
    across(everything(),
           ~ ifelse(. == "" | is.na(.), NA_character_, .))
  )
```

## Save load data

```{r}
write_tsv(
  dat_load, 
  here("data", "01_dat_load.tsv"), 
  na = "NA")
```
