==> R/16_analysis_12.qmd <==
---
title: "16_analysis_12"
format: 
  html:
    self-contained: true
---

## Load Libraries

```{r}
#| message: false
library("here")
library("tidyverse")
source(here("R/99_proj_func.R"))
```

## Load Data

```{r}
#| message: false
input_file <- here("data/03_dat_aug.tsv")
dat_aug <- read_tsv(
  input_file,
  na = c("", "NA"),
  show_col_types = FALSE)
```

## Plot

```{r}
gg_1 <- ggplot(
  dat_aug,
  aes(
    x = her2_status,
    y = her2_fish,
    fill = her2_status)) +
  geom_boxplot() +
  theme_minimal() +
  labs(
    title = "HER2-FISH and HER2 receptor Status Comparison",
    x = "HER2 Receptor",
    y = "HER2-FISH") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5))

gg_1
```

```{r}
gg_2 <- ggplot(
  dat_aug,
  aes(
    x = as.numeric(tumor_size_before), 
    y = her2_fish)) +
  geom_jitter(width = 0.2, 
              alpha = 0.5) +
  geom_smooth(method = "lm", 
              se = TRUE, 
              color = "blue") +
  theme_minimal() +
  labs(
    x = "Tumorsize",
    y = "HER2-FISH",
    title = "Linear Regression: HER2-FISH vs Tumorgsize") +
  theme(plot.title = element_text(hjust = 0.5))

gg_2
```

## Save plots

```{r}
save_plot(gg_1, 
          "16_plot_12_1.png")
save_plot(gg_2, 
          "16_plot_12_2.png")
```
