==> R/17_analysis_13.qmd <==
---
title: "17_analysis_13"
format: 
  html:
    self-contained: true
---

## Load Libraries

```{r}
#| message: false
library("here")
library("tidyverse")
source(here("R/99_proj_func.R"))
```

## Load Data

```{r}
#| message: false
input_file <- here("data/03_dat_aug.tsv")
dat_aug <- read_tsv(
  input_file,
  na = c("", "NA"),
  show_col_types = FALSE)
```

## PCA

### Set up the dataset

```{r}
df_clean <- dat_aug |>
  select(
    her2_num, 
    er_num,
    pr_num, 
    age) |>
  drop_na()

df_scaled <- df_clean |>
  map_dfc(~ (. - mean(.)) / sd(.))

pca_res <- prcomp(df_scaled, 
                  center = FALSE, 
                  scale. = FALSE)

pca_df <- as_tibble(pca_res$x) |>
  bind_cols(df_clean)
```

### Plot PCAs

```{r}
gg_1 <- pca_df |>
  ggplot(aes(
    x = PC1,
    y = PC2,
    color = factor(er_num))) +
  geom_point(size = 3,
             alpha = 0.7) +
  labs(
    title = "PCA Plot of ER Status",
    color = "ER Status") +
  theme_minimal() + 
  theme(plot.title = element_text(hjust = 0.5))

gg_2 <- pca_df |>
  ggplot(aes(
    x = PC1,
    y = PC2,
    color = factor(pr_num))) +
  geom_point(size = 3, 
             alpha = 0.7) +
  labs(
    title = "PCA Plot of PR Status",
    color = "PR Status") +
  theme_minimal() + 
  theme(plot.title = element_text(hjust = 0.5))

gg_3 <- pca_df |>
  ggplot(aes(
    x = PC1,
    y = PC2, 
    color = factor(her2_num))) +
  geom_point(size = 3, 
             alpha = 0.7) +
  labs(
    title = "PCA Plot of HER2 Status",
    color = "HER2 Status") +
  theme_minimal() + 
  theme(plot.title = element_text(hjust = 0.5))

gg_1
gg_2
gg_3
```

## Save plots

```{r}
save_plot(gg_1, 
          "17_plot_13_1.png")
save_plot(gg_2, 
          "17_plot_13_2.png")
save_plot(gg_3, 
          "17_plot_13_3.png")
```
