==> R/18_analysis_14.qmd <==
---
title: "18_analysis_14"
format: 
  html:
    self-contained: true
editor: visual
---

## Running Code

#### Loading libraries:

```{r}
#| message: false
library("here")
library("tidyverse")
library("scales")
source(here("R/99_proj_func.R"))
```

#### Read data:

```{r}
#| message: false
dataset <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

#### Re-ordering

Ordering the different genotypes into groups from the augmented data:

```{r}

# Create ER/PR/HER2 combo groups
dataset <- dataset |>
  mutate(
    ER = if_else(
      er_status == "P", 
      "ER+",
      "ER-"),
    PR = if_else(
      pr_status == "P",
      "PR+", 
      "PR-"),
    HER2 = if_else(
      her2_status == "P",
      "HER2+",
      "HER2-"),
    combo = paste(
      ER,
      PR,
      HER2,
      sep = ", ")) |>
  mutate(
    combo = factor(
      combo,
      levels = c(
        "ER+, PR-, HER2-",
        "ER+, PR+, HER2-",
        "ER+, PR-, HER2+",
        "ER+, PR+, HER2+",
        "ER-, PR+, HER2-",
        "ER-, PR+, HER2+",
        "ER-, PR-, HER2+",
        "ER-, PR-, HER2-")))

# check: how many rows total?
nrow(dataset)

# check: how many patients in each genotype?
dataset |> 
  count(combo)
```

#### **Graph 1:**

##### **Response (pCR vs RD) by ER/PR/HER2 Gentypes**

**pCR** = **pathological complete response,** meaning no residual invasive cancer found in the breast or lymph nodes at the time of surgery, It means the treatment eliminated all detectable invasive tumor. This is a strong **positive** prognostic indicator.

**RD = residual disease, meaning** invasive tumor remains in the breast and/or lymph nodes after treatment. It means the cancer did not completely respond to treatment. Indicates relative treatment resistance. RD is considered a **non-responder** outcome.

```{r}

# total number of patients
N_total <- nrow(dataset)

# Count N per genotype (combo) group
group_counts_resp <- dataset |>
  group_by(combo) |>
  summarise(N_group = n(),
            .groups = "drop")

custom_palette <- c("#e44b8d",
                    "#90C84C")

gg_14 <- ggplot(
  dataset, 
  aes(
    x = combo,
    fill = response)) +
  geom_bar(position = "fill") +
  scale_y_continuous(labels = percent_format()) +
  
  # show all 8 levels even if some are empty
  scale_x_discrete(drop = FALSE) +
  
  # add N labels above each bar
  geom_text(
    data = group_counts_resp,
    aes(
      x = combo,
      y = 1.05,
      label = paste0("N=", N_group)),
    size = 3,
    inherit.aes = FALSE) +
  labs(
    title = "Response (pCR=good vs RD=bad) by ER/PR/HER2 Genotypes",
    x = "ER/PR/HER2 Genotypes",
    y = paste0("Proportion of Patients (N = ", N_total, ")"),
    fill = "Response") +
  scale_fill_manual(values = custom_palette) +
  theme_minimal() +
  theme(axis.text.x  = element_text(angle = 45,
                                    hjust = 1,
                                    size = 8),
        plot.title   = element_text(hjust = 0.5),
        axis.title.x = element_text(vjust = 0)) +
  coord_cartesian(ylim = c(0, 1.1))

gg_14
```

##### Save plot

```{r}
save_plot(gg_14,
          "18_key_plot_14.png")
```

##### **Conclusions:**

Estrogen receptor-positive breast cancer patients have a worse treatment response (Graph 1).
