==> R/22_analysis_18.qmd <==
---
title: "22_analysis_18"
format: 
  html:
    self-contained: true
editor: visual
---

## Running Code

#### Loading libraries:

```{r}
#| message: false
library("here")
library("tidyverse")
library("scales")
source(here("R/99_proj_func.R"))
```

#### Read data:

```{r}
#| message: false
dataset <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

#### Re-ordering

Ordering the different genotypes into groups from the augmented data:

```{r}
# Create ER/PR/HER2 combo groups
dataset <- dataset |>
  mutate(
    ER = if_else(
      er_status == "P", 
      "ER+",
      "ER-"),
    PR = if_else(
      pr_status == "P",
      "PR+", 
      "PR-"),
    HER2 = if_else(
      her2_status == "P",
      "HER2+",
      "HER2-"),
    combo = paste(
      ER,
      PR,
      HER2,
      sep = ", ")) |>
  mutate(
    combo = factor(
      combo,
      levels = c(
        "ER+, PR-, HER2-",
        "ER+, PR+, HER2-",
        "ER+, PR-, HER2+",
        "ER+, PR+, HER2+",
        "ER-, PR+, HER2-",
        "ER-, PR+, HER2+",
        "ER-, PR-, HER2+",
        "ER-, PR-, HER2-")))

# check: how many rows total?
nrow(dataset)

# check: how many patients in each genotype?
dataset |> 
  count(combo)
```

#### **Graph 5:**

##### **Prep: treatment among genotypes**

```{r}
# Clean treatment variable
dat_treat <- dataset |>
  mutate(
    treatment = as.character(treatment),
    treatment = if_else(
      is.na(treatment),
      "NA",
      treatment))

# Summarise: count + proportions per genotype × treatment
treat_summary <- dat_treat |>
  group_by(combo,
           treatment) |>
  summarise(n = n(),
            .groups = "drop") |>
  group_by(combo) |>
  mutate(
    total = sum(n),
    pct   = n / total) |>
  ungroup()

# N per combo (for labels)
group_counts_treat <- treat_summary |>
  distinct(combo,
           total) |>
  rename(N_group = total)

# Keep the same genotype order you already defined
treat_summary$combo <- factor(
  treat_summary$combo,
  levels = levels(dataset$combo))
```

##### **Graph -** Treatment Distribution Across ER/PR/HER2 Genotypes

```{r}
# Construct color palett
treatment_colors <- c(
  "FAC"        = "#7f3667",
  "FACT"       = "#bb437e",
  "FACT+XRT/X" = "#e44b8d",
  "FEC"        = "#e2619f",
  "TFAC"       = "#e5a0c6",
  "FECT"       = "lightblue",
  "TFAC/HT"    = "#B3D84C",
  "TFEC"       = "#90C84C",
  "TH/FAC"     = "#70B83F",
  "TH/FEC"     = "#4F9A2F",
  "Tonly"      = "#1D6510",
  "TXFAC"      = "#004400",
  "NA"         = "grey")
```

```{r}
gg_18 <- ggplot(
  treat_summary,
  aes(
    x = combo,
    y = pct,
    fill = treatment)) +
  
  # N labels above bars 
  geom_text(
    data = group_counts_treat,
    aes(
      x = combo,
      y = 1.07,                     
      label = paste0("N=",
                     N_group)),
    size = 3,
    inherit.aes = FALSE) +
  
  # Stacked bars
  geom_col(position = "fill") +
  
  # Percent labels
  geom_text(
    aes(
      label = ifelse(
        pct > 0.02,
        paste0(round(pct * 100), "%"),
        "")),
    position = position_fill(vjust = 0.5),
    size = 2.5) +
  
  # Apply grey for NA + colors for other treatments
  scale_fill_manual(values = treatment_colors) +

  scale_x_discrete(drop = FALSE) +
  scale_y_continuous(labels = percent_format()) +
  labs(
    title = "Treatment Distribution Across ER/PR/HER2 Genotypes",
    x = "ER/PR/HER2 Genotype",
    y = "Proportion of Patients",
    fill = "Treatment") +
  theme_minimal() +
  theme(axis.text.x  = element_text(angle = 20,
                                    hjust = 1, 
                                    size = 8),
        plot.title   = element_text(hjust = 0.5),
        axis.title.x = element_text(vjust = 0),
        
        legend.position = "bottom",
        legend.title = element_text(size = 10),
        legend.text  = element_text(size = 8),
        legend.box.spacing = unit(0.2, "lines"),
        legend.key.size  = unit(3.5, "mm")) +
  guides(fill = guide_legend(
    ncol = 7,
    byrow = TRUE,
    title.position = "top")) +
  coord_cartesian(ylim = c(0, 1.1))

gg_18
```

##### Save plot

```{r}
save_plot(gg_18, 
          "22_key_plot_18.png")
```

##### Conclusion:

Generally, more different treatments were tried in the last two groups, especially for the second last group, which also had the best response. The second group from the left (ER+, PR+, HER2-) also had more treatments but showed a bad response. So in conclusion, (1) more personalized treatment options give better response, (2) and a general broad TFAC and TFEC is not working for estrogen-positive patients.

Aggressiveness, tumor size, and nodes do not necessarily mean a bad outcome. With personalized treatments, it can mean good results for the patient even if the stage or grade of the tumor seems to have a theoretically bad prognosis.
